feat: [WLEO-340] Add x5chain parsing (#12)
## Short description
This PR adds the `x5chain` parsing in the `unprotectedHeader` of a
`issuerAuth` CBOR document.

## List of changes proposed in this pull request
- Update native dependencies for iOS and Android;
- Update the Zod parser with the `x5chain` cliam.

## How to test
Test the `Decode issuer signed from single doc with decoded issuer auth`
scenario and check the console log result. The parsed document should
contain the `x5chain` claim under `unprotectedHeader` which is an array
of strings.